Parse nutritional data and split functionality into multiple files
